chg int to long int, in pointer cast, to help build on 64-bit Linux
git-svn-id: http://vtp.googlecode.com/svn/trunk@6162 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtui/ProjectionDlg.cpp
@@ -491,7 +491,7 @@ void ProjectionDlg::OnZone( wxCommandEvent &event )
 	TransferDataFromWindow();
 
 	void *vval = m_pZoneCtrl->GetClientData(m_iZone);
-	int val = (int) vval - 100;
+	long int val = (long int) vval - 100;
 	m_proj.SetUTMZone(val);
 
 	UpdateControlStatus();
